+++ ReleaseNotes/release_notes.md
@@ -33,6 +33,7 @@ minuit.release ('p3')
   1. Add parameter `fit_fun` for  `toys` - that allows to (re)define the default "fit"-policy 
   1. Add parameter `gen_fun` for  `toys` - that allows to (re)define the default "generation"-policy 
   1. `test_plotting_summary_graph.py` : add call for `ROOT.gPad.RedrawAxis` -  thanks to Tom Blake 
+  1. add `**kwargs` for all `parallel`-methods, arguments are used for `WorkManager`
 
 ## Backward incompatible changes: